requirements.txt
streamlit>=1.32.0
mistralai>=0.0.12
ipython>=8.12.0
python-dotenv>=1.0.0 